+++ app/preload.cjs
@@ -5,7 +5,7 @@ contextBridge.exposeInMainWorld('dePerPortal',{
   action:data=>ipcRenderer.invoke('action',data),
   select:data=>ipcRenderer.invoke('select',data),
   game:value=>ipcRenderer.invoke('game',value),
-  clearTraps:()=>ipcRenderer.invoke('clear-traps'),
+  resetTrapDetections:()=>ipcRenderer.invoke('reset-trap-detections'),
   restoreTrapBackup:()=>ipcRenderer.invoke('restore-trap-backup'),
   rescan:()=>ipcRenderer.invoke('rescan'),
   launch:()=>ipcRenderer.invoke('launch'),